requirements.txt
langchain
langgraph
langchain-community
langchain-core
langchain-openai
langchain-groq
groq
faiss-cpu
sentence-transformers
typing
uuid
python-dotenv
langchain-huggingface
langgraph-checkpoint-sqlite
transformers
requests
pypdf
pydantic
pydantic-ai